Min nucleotide from a range of sequence DNA
A DNA sequence can be represented as a string consisting of the letters A, C, G and T which have impact factors of 1, 2, 3 and 4, respectively
There are M queries, which are given in non-empty arrays P and Q, each consisting of M integers. The K-th query (0 ≤ K < M) requires you to find the minimal impact factor of nucleotides contained in the DNA sequence between positions P[K] and Q[K] (inclusive).